+++ src/config/settings.py
@@ -126,7 +126,7 @@ DATABASES = {
         "NAME": env("DATABASE", default="email_archive_django"),
         "USER": env("DATABASE_USER", default="user"),
         "PASSWORD": env("DATABASE_PASSWORD", default="passwd"),
-        "HOST": "db",
+        "HOST": env("DATABASE_HOST", default="db"),
         "PORT": env("DATABASE_PORT", default=""),
         "OPTIONS": (
             {